Add email field to the login screen and change 'OnSignin', 'OnSignup' from a react function to vanilla JS function
User name field was added because when the user creates  an account the email does not need to be unique but username must be. Multiple accounts can be connected to a single email address. Also, changed react functions to vanilla JS because I am no longer using hooks for api calls